chore: Restructure folders * group components into module folders
