+++ src/app/app.module.ts
@@ -4,9 +4,9 @@ import { FormsModule } from "@angular/forms";
 import { RouterModule } from "@angular/router";
 import { HttpClientModule } from "@angular/common/http";
 import { AppComponent } from "./app.component";
-import { HomeComponent } from "./components/home/home.component";
-import { UserModule } from "./components/users/user.module";
-import { PostModule } from "./components/posts/post.module";
+import { HomeComponent } from "./home/home.component";
+import { UserModule } from "./users/user.module";
+import { PostModule } from "./posts/post.module";
 import { AppRoutingModule } from "./app-routing.module";
 
 @NgModule({
