+++ src/app/posts/post.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
-import { PostsComponent } from "./posts.component";
+import { PostsComponent } from "./post-list/post-list.component";
 import { RouterModule } from "@angular/router";
 
 @NgModule({
