+++ src/app/posts/post.service.spec.ts
@@ -1,11 +1,11 @@
-import { TestBed } from '@angular/core/testing';
+import { TestBed } from "@angular/core/testing";
 
-import { PostService } from './post.service';
+import { PostService } from "./post.service";
 
-describe('PostService', () => {
+describe("PostService", () => {
   beforeEach(() => TestBed.configureTestingModule({}));
 
-  it('should be created', () => {
+  it("should be created", () => {
     const service: PostService = TestBed.get(PostService);
     expect(service).toBeTruthy();
   });
