+++ src/app/users/user.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
 import { RouterModule } from "@angular/router";
-import { UsersComponent } from "./users.component";
+import { UsersComponent } from "./user-list/user-list.component";
 import { UserDetailComponent } from "./user-detail/user-detail.component";
 
 @NgModule({
